refactor: change folder structure in onboarding folder
