+++ lib/presentation/onboarding/preparation_name_select/cubit/preparation_name/preparation_name_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:on_time_front/presentation/onboarding/cubit/preparateion_step_name/preparation_step_name_cubit.dart';
+import 'package:on_time_front/presentation/onboarding/preparation_name_select/cubit/preparation_step_name/preparation_step_name_cubit.dart';
 
 part 'preparation_name_state.dart';
